Vista y funcionalidad de factura con total dinámico

+++ database/migrations/2021_01_11_192355_detail_bill.php
@@ -23,8 +23,8 @@ class DetailBill extends Migration
             $table->foreign('product_id')
             ->references('id')->on('products');
             $table->integer('quantity');
-            $table->float('unit_price');
-            $table->decimal('subtotal');
+            $table->integer('unit_price');
+            $table->integer('subtotal');
             $table->timestamps();           
         });
